requirements.txt
asgiref==3.8.1
Django==5.0.7
PyMySQL==1.1.1
sqlparse==0.5.1
djangorestframework~=3.15.2